Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    int a,b;
+    float a,b;
     float c;
-    scanf("%d",&a);
-    scanf("%d",&a);
+    scanf("%f",&a);
+    scanf("%f",&a);
     c=(a*b)/2;
     printf("Average: %.2f",c);
     return 0;
